move DiskIdentity
