+++ key-manager/Cargo.toml
@@ -14,6 +14,7 @@ key-manager-types.workspace = true
 omicron-common.workspace = true
 secrecy.workspace = true
 sha3.workspace = true
+sled-agent-types.workspace = true
 slog.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
