+++ nexus/db-model/src/deployment.rs
@@ -63,13 +63,13 @@ use nexus_types::deployment::{
 };
 use omicron_common::address::Ipv6Subnet;
 use omicron_common::address::SLED_PREFIX_LENGTH;
-use omicron_common::disk::DiskIdentity;
 use omicron_common::zpool_name::ZpoolName;
 use omicron_uuid_kinds::{
     BlueprintKind, BlueprintUuid, DatasetKind, ExternalIpKind, ExternalIpUuid,
     GenericUuid, MupdateOverrideKind, OmicronZoneKind, OmicronZoneUuid,
     PhysicalDiskKind, SledKind, SledUuid, ZpoolKind, ZpoolUuid,
 };
+use sled_agent_types::disk::DiskIdentity;
 use sled_agent_types::inventory::NetworkInterface;
 use sled_agent_types::inventory::OmicronZoneDataset;
 use sled_agent_types::inventory::SourceNatConfigGeneric;
