+++ nexus/db-queries/src/db/datastore/deployment.rs
@@ -3178,7 +3178,6 @@ mod tests {
     use omicron_common::api::internal::shared::PrivateIpConfig;
     use omicron_common::api::internal::shared::PrivateIpv4Config;
     use omicron_common::api::internal::shared::PrivateIpv6Config;
-    use omicron_common::disk::DiskIdentity;
     use omicron_common::zpool_name::ZpoolName;
     use omicron_test_utils::dev;
     use omicron_test_utils::dev::poll::CondCheckError;
@@ -3190,10 +3189,11 @@ mod tests {
     use omicron_uuid_kinds::ZpoolUuid;
     use pretty_assertions::assert_eq;
     use rand::Rng;
+    use sled_agent_types::disk::DiskIdentity;
+    use sled_agent_types::disk::M2Slot;
     use sled_agent_types::inventory::NetworkInterface;
     use sled_agent_types::inventory::NetworkInterfaceKind;
     use std::collections::BTreeMap;
-    use sled_agent_types::disk::M2Slot;
     use std::collections::BTreeSet;
     use std::mem;
     use std::net::Ipv6Addr;
