+++ nexus/test-utils/src/resource_helpers.rs
@@ -77,7 +77,6 @@ use omicron_common::api::external::UserId;
 use omicron_common::disk::DatasetConfig;
 use omicron_common::disk::DatasetKind;
 use omicron_common::disk::DatasetName;
-use omicron_common::disk::DiskIdentity;
 use omicron_common::disk::SharedDatasetConfig;
 use omicron_common::zpool_name::ZpoolName;
 use omicron_sled_agent::sim::SledAgent;
@@ -92,6 +91,7 @@ use omicron_uuid_kinds::ZpoolUuid;
 use oxnet::IpNet;
 use oxnet::Ipv4Net;
 use oxnet::Ipv6Net;
+use sled_agent_types::disk::DiskIdentity;
 use sled_agent_types::disk::OmicronPhysicalDiskConfig;
 use sled_agent_types::inventory::ZpoolHealth;
 use slog::debug;
