+++ sled-agent/config-reconciler/src/host_phase_2.rs
@@ -723,13 +723,13 @@ mod tests {
     use assert_matches::assert_matches;
     use bytes::BufMut as _;
     use camino_tempfile::Utf8TempDir;
-    use omicron_common::disk::DiskIdentity;
     use omicron_test_utils::dev;
     use omicron_uuid_kinds::InternalZpoolUuid;
     use proptest::collection::vec;
     use proptest::prelude::*;
     use proptest::prop_oneof;
     use sha2::Digest as _;
+    use sled_agent_types::disk::DiskIdentity;
     use sled_storage::config::MountConfig;
     use slog_error_chain::InlineErrorChain;
     use std::io::Write;
