+++ sled-agent/config-reconciler/src/internal_disks.rs
@@ -18,9 +18,9 @@ use iddqd::IdOrdMap;
 use iddqd::id_upcast;
 use omicron_common::backoff::Backoff as _;
 use omicron_common::backoff::ExponentialBackoffBuilder;
-use omicron_common::disk::DiskIdentity;
 use omicron_common::zpool_name::ZpoolName;
 use omicron_uuid_kinds::InternalZpoolUuid;
+use sled_agent_types::disk::DiskIdentity;
 use sled_agent_types::disk::DiskVariant;
 use sled_agent_types::disk::M2Slot;
 use sled_hardware::PooledDiskError;
