+++ sled-agent/config-reconciler/src/raw_disks.rs
@@ -142,12 +142,12 @@ fn can_remove_disk(disk: &RawDisk, log: &Logger) -> bool {
 mod tests {
     use super::*;
     use camino::Utf8PathBuf;
-    use omicron_common::disk::DiskIdentity;
     use omicron_test_utils::dev;
     use proptest::collection::btree_map;
     use proptest::prelude::*;
     use proptest::sample::Index;
     use proptest::sample::size_range;
+    use sled_agent_types::disk::DiskIdentity;
     use sled_agent_types::disk::DiskVariant;
     use sled_hardware::DiskFirmware;
     use sled_hardware::UnparsedDisk;
