+++ sled-agent/rack-setup/src/plan/service.rs
@@ -1372,6 +1372,7 @@ mod tests {
     use omicron_common::api::internal::shared::AllowedSourceIps;
     use omicron_test_utils::dev::test_setup_log;
     use oxnet::Ipv6Net;
+    use sled_agent_types::disk::DiskIdentity;
     use sled_agent_types::early_networking::PortConfig;
     use sled_agent_types::early_networking::RackNetworkConfig;
     use sled_agent_types::early_networking::UplinkPorts;
@@ -1551,7 +1552,7 @@ mod tests {
 
         let disks: Vec<_> = (0..DISK_COUNT)
             .map(|i| sled_agent_types::inventory::InventoryDisk {
-                identity: omicron_common::disk::DiskIdentity {
+                identity: DiskIdentity {
                     vendor: "vendor".to_string(),
                     model: "model".to_string(),
                     serial: format!("test-{i}"),
