+++ sled-agent/resolvable-files/src/test_utils.rs
@@ -5,10 +5,10 @@
 use std::sync::Arc;
 
 use camino::Utf8Path;
-use omicron_common::disk::DiskIdentity;
 use omicron_uuid_kinds::InternalZpoolUuid;
 use sled_agent_config_reconciler::InternalDiskDetails;
 use sled_agent_config_reconciler::InternalDisksReceiver;
+use sled_agent_types::disk::DiskIdentity;
 use sled_storage::config::MountConfig;
 
 pub(crate) fn make_internal_disks_rx(
