+++ sled-agent/types/versions/src/latest.rs
@@ -46,6 +46,7 @@ pub mod diagnostics {
 
 pub mod disk {
     pub use crate::v1::disk::DiskEnsureBody;
+    pub use crate::v1::disk::DiskIdentity;
     pub use crate::v1::disk::DiskPathParam;
     pub use crate::v1::disk::DiskStateRequested;
     pub use crate::v1::disk::DiskVariant;
