+++ sled-hardware/src/illumos/mod.rs
@@ -12,7 +12,7 @@ use camino::Utf8PathBuf;
 use gethostname::gethostname;
 use illumos_devinfo::{DevInfo, DevLinkType, DevLinks, Node, Property};
 use libnvme::{Nvme, controller::Controller};
-use omicron_common::disk::DiskIdentity;
+use sled_agent_types::disk::DiskIdentity;
 use sled_agent_types::disk::DiskVariant;
 use sled_hardware_types::{Baseboard, OxideSled, SledCpuFamily};
 use slog::Logger;
