+++ sled-storage/src/dataset.rs
@@ -14,9 +14,8 @@ use illumos_utils::zpool::ZpoolName;
 use key_manager::StorageKeyRequester;
 use omicron_common::api::external::ByteCount;
 use omicron_common::api::internal::shared::DatasetKind;
-use omicron_common::disk::{
-    CompressionAlgorithm, DatasetName, DiskIdentity, GzipLevel,
-};
+use omicron_common::disk::{CompressionAlgorithm, DatasetName, GzipLevel};
+use sled_agent_types::disk::DiskIdentity;
 use sled_agent_types::disk::DiskVariant;
 use slog::{Logger, debug, info, warn};
 use slog_error_chain::InlineErrorChain;
