+++ test-utils/src/dev/mod.rs
@@ -22,9 +22,9 @@ use dropshot::ConfigLogging;
 use dropshot::ConfigLoggingIfExists;
 use dropshot::ConfigLoggingLevel;
 pub use dropshot::test_util::LogContext;
-use omicron_common::disk::DiskIdentity;
 use sha2::Digest;
 use sha2::Sha256;
+use sled_agent_types::disk::DiskIdentity;
 use slog::Logger;
 use std::io::BufReader;
 use std::io::Read;
